[tau] Move tau-agent to a proper Python package
Move top-level modules into a tau/ package:
- tau.py -> tau/app.py (with main() entry point)
- tools.py -> tau/tools.py
- session.py -> tau/session.py

Add tau/__init__.py and tau/__main__.py so the app can be invoked as
`python -m tau` or via the `tau` script entry point.

Add build-system (hatchling) and project.scripts to pyproject.toml.
