+++ examples/tau-agent/pyproject.toml
@@ -8,6 +8,16 @@ dependencies = [
     "textual>=3.0",
 ]
 
+[project.scripts]
+tau = "tau.app:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["tau"]
+
 [tool.uv.sources]
 ai = { path = "../..", editable = true }
