+++ examples/tau-agent/tau/__init__.py
@@ -1,0 +1,1 @@
+"""tau — a coding-agent chat bot built on the `ai` library and Textual."""
