+++ examples/tau-agent/tau/__main__.py
@@ -1,0 +1,5 @@
+"""Entry point for ``python -m tau``."""
+
+from tau.app import main
+
+main()
